Allow to enter story with leading '#' (e.g: #1234)
An easy way to get a story ID from Pivotal Tracker is to press the
button with the text "ID" on the story form of Tracker's Web UI. The
problem with that is that it includes the leading '#' (e.g. #101899896)
and this plugin currently fails if you include it on the "Get Story"
prompt. So one has to press paste the story ID on the clipboard and then
go to the start of the prompt text and remove the '#'.

Allow to optionally include the trailing '#' in order to make it easier
to just copy/paste the story's ID.

Achieve this by removing the trailing '#' if it is present.