Fix exception on empty preference file
Fix exception that gets thrown when the preference file is empty or a
certain property in that file doesn't exist. This exception is usually
thrown when sublime is booting, during the process of loading packages,
done by Package Control

This could be causing the plugin to be disabled after restarting
sublime for some users.